requirements.txt
numpy==1.9.2
pyOSC==0.3.5b-5294
wsgiref==0.1.2
